+++ fe/fe-core/src/main/java/com/starrocks/sql/plan/ScalarOperatorToExpr.java
@@ -125,7 +125,7 @@ public class ScalarOperatorToExpr {
 
             if (node.getName() == "lambda") {
                 Expr col = new SlotRef(new TableName("lambda", "lambda"), "lambda");
-                col.setType(Type.INT);
+                col.setType(Type.TINYINT);
                 return col;
             }
